Add image augmentation utility and testing framework
- Implemented `augment_image` function in `augmentation.py` for applying various transformations to images, including rotation, shearing, resizing, grayscale conversion, brightness and contrast adjustments, Gaussian blur, noise injection, and adding distractor lines.
- Created `augmentations_explained.md` documentation detailing the augmentation process, parameters, and usage examples.
- Added `test_augmentation.py` script to visualize the effects of individual augmentations on a sample image, generating a grid of augmented images.
- Included a benchmark function to measure the performance of the augmentation process.
- Added a sample test image `test.png` for demonstration purposes.
