+++ outputs/cnn_part3_augmentated/training_log.txt
@@ -1,0 +1,10 @@
+[2025-09-03 18:52:40] 
+[2025-09-03 18:52:40] ========================================
+[2025-09-03 18:52:40] Optuna Tuning Started from configs/part3_augmentation_best_cnn.yaml
+[2025-09-03 18:52:40] ========================================
+[2025-09-03 18:52:40] 
+[2025-09-03 18:52:41] 
+[2025-09-03 18:52:41] ========================================
+[2025-09-03 18:52:41] Scenario: aug_to_mirror_part_3
+[2025-09-03 18:52:41] ========================================
+[2025-09-03 18:52:41]
